Change capture divider from uint16 to unt8

+++ platformio/src/acquisition/analyzer.h
@@ -60,7 +60,7 @@ struct AdcCaptureBuffer {
   // Indcates the X time divider >= 1. Value of 1 indicates all
   // samples are included. Value of 2 indicates every other sample
   // is included and so on.
-  uint16_t divider;
+  uint8_t divider;
 
   // True if capture was synced with a trigger event where
   // v1 crossed up the y=0 axis. The triggered event, if available
@@ -236,14 +236,14 @@ void set_is_reversed_direction(bool is_reverse_direction);
 bool get_is_reversed_direction();
 
 // Clipped internally to allowed range.
-void set_signal_capture_divider(uint16_t divider);
+void set_signal_capture_divider(uint8_t divider);
 
 // Return a copy of the internal settings. Used after
 // calibrate_zeros() to save the current settings in the
 // EEPROM.
 void get_settings(Settings* settings);
 
-// @@@ Temp
+// Temp
 void dump_dma_state();
 
 }  // namespace analyzer
